# replace bytes.Buffer with strings.Builder