strings.Builder has better performance over bytes.Buffer for building
strings:

Benchmark results:
```
name old time/opnew time/opdelta
RequirementString-892.6ns ± 1%46.9ns ± 1%  -49.34%  (p=0.000 n=4+5)

name old alloc/op   new alloc/op   delta
RequirementString-8 96.0B ± 0% 32.0B ± 0%  -66.67%  (p=0.008 n=5+5)

name old allocs/op  new allocs/op  delta
RequirementString-8  2.00 ± 0%  1.00 ± 0%  -50.00%  (p=0.008 n=5+5)
```

```
name  old time/opnew time/opdelta
_ParseQuotedString-8 146ns ±20% 105ns ± 2%  -28.14%  (p=0.008 n=5+5)

name  old alloc/op   new alloc/op   delta
_ParseQuotedString-8 80.0B ± 0% 24.0B ± 0%  -70.00%  (p=0.008 n=5+5)

name  old allocs/op  new allocs/op  delta
_ParseQuotedString-8  2.00 ± 0%  2.00 ± 0% ~ (all equal)
```

Signed-off-by: André Martins <aanm90@gmail.com>

**What type of PR is this?**

/kind cleanup

```release-note
NONE
```