@pacoxu 2020-12-28
Smart. It inspired me a similar docker pr https://github.com/moby/moby/pull/40856/files which fixed conditional memory leak.