@thockin 2020-12-28
pkg/proxy/iptables/proxier.go makes heavy use of bytes.Buffer - maybe
your next target? :)

On Mon, Dec 28, 2020 at 8:33 AM André Martins <notifications@github.com> wrote:
>
> /retest
>
> —
> You are receiving this because your review was requested.
> Reply to this email directly, view it on GitHub, or unsubscribe.