@aanm 2020-12-28
> pkg/proxy/iptables/proxier.go makes heavy use of bytes.Buffer - maybe your next target? :)
> […](#)
> On Mon, Dec 28, 2020 at 8:33 AM André Martins ***@***.***> wrote: /retest — You are receiving this because your review was requested. Reply to this email directly, view it on GitHub, or unsubscribe.

Thanks for the suggestion @thockin. Although, I think it will look bad if I make changes in the `pkg/proxy/iptables` given that people can simply use Cilium with eBPF :see_no_evil: