@caesarxuchao 2021-01-15
/lgtm
/approve

It's interesting that the strings.Builder.WriteByte(), which simply does `b.buf = append(b.buf, c)` has better performance than bytes.Buffer.WriteByte().